/me serves the zone the seat was stored with
The round trip was built at both ends and severed in the middle. Signup captures
the browser's zone, values.ParseTimezone validates it, and the row stores it on
the workspace and the first user — and meResponse then built the User without
the field, so the endpoint every client fetches first never said it.

Nothing downstream could localize an instant to the person reading it. #26's own
words are that everything else in it is blocked on this, and the frontend's
twelve hard-coded "Europe/Berlin" literals are what a caller writes when no
correct answer is served.

The zone rides the session read beside the locale, for the reason the locale
already gives there: the SPA needs to know which catalog to render before it
draws anything, and which zone to render an instant in before it prints a date.
A second round-trip for two columns would paint the wrong language and the wrong
day first.

Omitted rather than defaulted when empty. app_user.timezone is NOT NULL with
'UTC', so in practice the field is always there and this is a guard on the
SHAPE: Identity is an ordinary struct several paths build partially, and if one
reached this response the field would arrive as "" — which a client parses as a
zone and fails on. Omitted says "ask the browser"; empty says "here is a zone"
and lies.

Two files went over the 500-LOC cap and are split by concept rather than
trimmed. handlers.go held both the HTTP verbs and the projection they answer
with — a request's cookies and statuses is a different subject from one struct's
shape on the wire. service.go held both the session lifecycle and the grant
resolution its three entry points each end by calling — proving who is asking is
a different subject from what the answer entitles them to.

Both assertions mutation-checked: dropping the field reproduces the original
defect exactly, and serving &id.Timezone instead of omitting fails the
empty-string guard.

Part of #26 — the blocking finding only. The zone-by-purpose helper, the
workspace reporting zone on a GET, the frontend useZone() seam and the literals
are the rest of that epic.
